+++ flashfold/scripts/__init__.py
@@ -4,4 +4,5 @@ from .download_sequence_db import *
 from .predict_structure import *
 from .download_ncbi_assembly import *
 from .create_json_with_ligand import *
+from .run_alphafold3 import *
 from .generate_report import *